requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9     
pymysql==1.1.1             
python-dotenv==1.0.1